Change the symbolic manipulation in the project from GiNaC to the lower-level SymEngine. This means that the license can be changed from GPLv3 to BSD. It is also to prepare for later linking against libraries which in turn use SymEngine.
There are some differences in the generated mu polynomials. This appears to be due to poor conditioning and different rounding. For example, an extreme case, for mu<8,6> the third coefficient differ already in the 6th digit. However, the coefficient is ~10^-25 and the coefficients it is computed from is up to ~10^-11 so we are looking at ~10^-16 relative errors so it seems it has something to do with double rounding. Performing the computations manually with the original mpreal objects, it seems it is now correct and the previous version had rounding issues. However, I cannot observe any difference when it comes to the actual transformation errors.